+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Changes in version ?.?.? - 2016-??-??
+
+ * Medium changes
+   - Include the hostname in directory source entries of consensuses
+     and votes.
+
+
 # Changes in version 1.1.0 - 2015-12-28
 
  * Medium changes
@@ -26,8 +33,6 @@
    - Support hidden-service statistics in extra-info descriptors.
    - Support onion-key and ntor-onion-key cross certificates in server
      descriptors.
-   - Include the hostname in directory source entries of consensuses
-     and votes.
 
  * Minor changes
    - Start using Java 7 features like the diamond operator and switch